Compiler unification/gaps: jvm

+++ platforms/jvm/src/emitter/io_adapter.rs
@@ -25,7 +25,7 @@ fn set(chunk: &mut Chunk, slot: u16, line: u32) {
     chunk.emit_op_u16(Op::LOCAL_SET, slot, line);
 }
 
-fn key(chunk: &mut Chunk, name: &str) -> u16 {
+fn key(chunk: &mut Chunk, name: &str) -> u32 {
     chunk.add_constant(vybe_runtime::Value::String(name.into()))
 }
